New mothod = Create Hex M1 / Parse Encode Hex M2 A new mode added to create own asterix message of ours.
Create Hex M1
-------------
bool USE_SAMPLE_SCENARIO_CREATE_HEX_DATA{0}; is used for M1 which will take the radarMsgHandler.asterixMessageParser(vec, USE_SAMPLE_SCENARIO_CREATE_HEX_DATA);
the vec data will be parsed and encoded to create the hex data. If giving vec hex data is not same after parsing and encoding then serialization or parsing has an error.
vec.data() == encodedHex.data() // Which it will be printed to terminal after encoding as a string

bool USE_SAMPLE_SCENARIO_CREATE_HEX_DATA{1}; is taking a sample json and without parsing it will be directly go to encoding and create a Hex data.
Which you can use in your Mode 0 to check if the created hex data will be the same as after using parsing and encoding message.
Advantage of it ,we can manually create messages for asterix or manually change the value of message fields. And we can get the hex data for this json sample.